backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
openai==1.51.0
python-dotenv==1.0.0
httpx>=0.26.0,<0.28.0
PyJWT==2.8.0
requests==2.31.0
supabase==2.9.0

pillow==10.1.0
numpy<2.0.0
ultralytics>=8.0.0